extracted internal integration tests
